requirements.txt
requests
dnspython
pythonping